move number utils to localstack.utils.numbers
